requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
# mysql-connector-python==8.2.0  # Uncomment for MySQL production
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dateutil==2.8.2

mysql-connector-python==8.2.0

# Testing dependencies (optional - for running test_complete_api.py)
aiohttp==3.9.1
psutil==5.9.6
requests==2.31.0 